refactor(assets): change directory
