+++ projects/common-components/src/lib/common-components.module.ts
@@ -5,7 +5,7 @@ import { ButtonComponent } from "./button/button.component";
 import { TalksListComponent } from "./talks/talks-list/talks-list.component";
 import { TalkItemComponent } from "./talks/talk-item/talk-item.component";
 import { TalkItemsLineComponent } from "./talks/talk-items-line/talk-items-line.component";
-import { SortPipe } from "./pipes/sort/sort.pipe";
+import { SortPipe } from "@jsmu/shared";
 
 @NgModule({
 	declarations: [
