+++ projects/common-components/src/lib/talks/talk-item/talk-item.component.html
@@ -7,11 +7,7 @@
 	<div class="jsmu-talk-item__wrapper">
 		<div class="jsmu-talk-item__photo">
 			<a href="#">
-				<img
-					class="jsmu-talk-item__photo-img"
-					src="../../../assets/images/cat.jpg"
-					alt="profile"
-				/>
+				<img class="jsmu-talk-item__photo-img" src="/assets/images/cat.jpg" alt="profile" />
 			</a>
 		</div>
 		<div class="jsmu-talk-item__content">
