+++ projects/shared/src/public-api.ts
@@ -1,7 +1,4 @@
-/*
- * Public API Surface of shared
- */
-
 export * from "./lib/shared.service";
 export * from "./lib/shared.component";
 export * from "./lib/shared.module";
+export * from "./lib/pipes/sort/sort.pipe";
